% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{datafiles}
\alias{datafiles}
\title{List of a few datasets commonly used in our workshops.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Data}{Colloquial name of data}
  \item{Remote}{URL of file}
  \item{Local}{Name to use for file on disc}
  \item{Type}{Type of file, such as CSV, Excel}
  \item{Mode}{Way to write file to disc such as 'w' for text files or 'wb' for binary files}
  \item{Tag}{Tag for the type of data}
  \item{Purpose}{What the data is used to show}
}}
\usage{
datafiles
}
\description{
A dataset containing information about data to download for a workshop.
}
\keyword{datasets}
